+++ packages/theme/src/types.ts
@@ -82,7 +82,7 @@ export type BackstagePaletteAdditions = {
     error: string;
     text: string;
     link: string;
-    warning: string;
+    warning?: string;
   };
 };