+++ plugins/api-docs/src/components/OpenApiDefinitionWidget/index.ts
@@ -15,3 +15,4 @@
  */
 
 export { OpenApiDefinitionWidget } from './OpenApiDefinitionWidget';
+export type { OpenApiDefinitionWidgetProps } from './OpenApiDefinitionWidget';