+++ plugins/azure-devops/src/components/BuildTable/BuildTable.tsx
@@ -18,7 +18,8 @@ import { Box, Typography } from '@material-ui/core';
 import {
   BuildResult,
   BuildStatus,
-} from 'azure-devops-node-api/interfaces/BuildInterfaces';
+  RepoBuild,
+} from '@backstage/plugin-azure-devops-common';
 import {
   Link,
   ResponseErrorPanel,
@@ -34,7 +35,6 @@ import {
 
 import { DateTime } from 'luxon';
 import React from 'react';
-import { RepoBuild } from '../../api/types';
 
 export const getBuildResultComponent = (result: number | undefined) => {
   switch (result) {
@@ -149,8 +149,9 @@ const columns: TableColumn[] = [
     field: 'queueTime',
     width: 'auto',
     render: (row: Partial<RepoBuild>) =>
-      DateTime.fromISO(
-        row.queueTime ? row.queueTime.toString() : new Date().toString(),
+      (row.queueTime
+        ? DateTime.fromISO(row.queueTime)
+        : DateTime.now()
       ).toRelative(),
   },
 ];