+++ plugins/scaffolder-backend-module-rails/src/index.ts
@@ -15,7 +15,7 @@
  */
 
 /**
- * A module for the scaffolder backend that lets you template projects using Rails
+ * A module for the scaffolder backend that lets you template projects using {@link https://guides.rubyonrails.org/rails_application_templates.html | Rails}.
  *
  * @packageDocumentation
  */